pyproject.toml
[tool.poetry]
name = "padam"
version = "0.1.0"
description = "Python Aided Design and Manifacturing"
authors = ["Massimiliano Ravelli <massimiliano.ravelli@gmail.com>"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
solidpython = "^1.1.3"
pydantic = "^2.10.1"
setuptools = "^75.6.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.6.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
